Add updating algorithms for a Ising-spin system
The current interface to initialize each struct is developed still.  It
is not useful to receive an adjacency matrix from a user because all
structs accept any matrices for now.  One has to send a symmetric
matrix whose all diagonal components are zero.  To improve this
interface, we should make a restriction that a user passes either
upper or lower triangular parts of off-diagonal components.