Replace the procedure to generate a 2d-grid graph in the demonstration by that in NetworkX
Although I tried to generate such a graph by the Julia library
Graphs.jl, it does not seem a good way for now.  It is not easy to add
weights to nodes/edges of the graph.  Instead, I used PyCall.jl and the
Python library NetworkX.  To complete the current way only by use of
Julia, we have to modify Graphs.jl or to newly make a graph-generating
library.